requirements.txt
viser 
pupil-apriltags
loguru 
scipy
opencv-python
numpy
pyyaml
pyudev
termcolor